move service utils to its own directory
