+++ tsconfig.json
@@ -5,7 +5,6 @@
         "lib": ["ES2020", "DOM", "DOM.Iterable"],
         "module": "ESNext",
         "skipLibCheck": true,
-
         /* Bundler mode */
         "moduleResolution": "bundler",
         "allowImportingTsExtensions": true,
@@ -13,7 +12,6 @@
         "isolatedModules": true,
         "noEmit": true,
         "jsx": "react-jsx",
-
         /* Linting */
         "strict": true,
         "noUnusedLocals": true,
@@ -26,11 +24,16 @@
             "@components": ["./src/components/index"],
             "@pages": ["./src/pages/index"],
             "@services": ["./src/services/index"],
+            "@services/utils": ["./src/services/utils/index"],
             "@providers": ["./src/providers/index"],
             "@utils": ["./src/utils/index"],
             "@mocks/providers": ["./src/providers/__mocks__"]
         }
     },
     "include": ["src"],
-    "references": [{ "path": "./tsconfig.node.json" }]
+    "references": [
+        {
+            "path": "./tsconfig.node.json"
+        }
+    ]
 }
